refactor: create new core::decoding module
