+++ src/core/decoding/mod.rs
@@ -1,0 +1,1 @@
+pub mod error;
