+++ src/core/reader/mod.rs
@@ -1,4 +1,4 @@
-use crate::core::error::DecodingError;
+use crate::core::decoding::error::DecodingError;
 use crate::core::reader::span::Span;
 
 pub mod section_header;
