+++ src/core/reader/section_header.rs
@@ -1,4 +1,4 @@
-use crate::core::error::DecodingError;
+use crate::core::decoding::error::DecodingError;
 use crate::core::reader::span::Span;
 use crate::core::reader::WasmReader;
 use crate::core::utils::ToUsizeExt;
