+++ src/core/reader/types/global.rs
@@ -1,4 +1,4 @@
-use crate::core::error::DecodingError;
+use crate::core::decoding::error::DecodingError;
 use crate::core::reader::span::Span;
 use crate::core::reader::types::ValType;
 use crate::core::reader::WasmReader;
